+++ src/store/modules/ADempiere/pointOfSales/orderLine/getters.js
@@ -1,0 +1,10 @@
+/**
+ * Order Line Getters
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+
+export default {
+  getListOrderLine: (state) => {
+    return state.listOrderLine
+  }
+}
