+++ src/store/modules/ADempiere/pointOfSales/orderLine/mutations.js
@@ -1,0 +1,11 @@
+
+/**
+ * Order Line Mutations
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+
+export default {
+  setListOrderLine(state, listOrderLine) {
+    state.listOrderLine = listOrderLine
+  }
+}
