+++ src/store/modules/ADempiere/pointOfSales/orderLine/state.js
@@ -1,0 +1,8 @@
+/**
+ * Order Line State
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+
+export default {
+  listOrderLine: []
+}
