+++ src/store/modules/ADempiere/pointOfSales/productPrice/index.js
@@ -1,0 +1,18 @@
+
+import state from './state.js'
+import mutations from './mutations.js'
+import actions from './actions.js'
+import getters from './getters.js'
+
+/**
+ * Product Price Vuex Module
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+const listProductPrice = {
+  state,
+  mutations,
+  actions,
+  getters
+}
+
+export default listProductPrice
